# crates/core/src/openapi.yaml
openapi: 3.1.0
info:
  title: CrossNotifier Server API
  version: "1.0.0"
  description: |
    HTTP API for the CrossNotifier server. Accepts notifications from
    senders and broadcasts them to connected WebSocket clients.

    Senders can optionally request a synchronous response by setting
    `wait` / `maxWait` on the notification — the server will hold the
    HTTP connection open until a client resolves the notification or
    the timeout elapses.
servers:
  - url: http://localhost:9876
    description: Local server (default port)
security:
  - bearerAuth: []
...
paths:
  /notify:
    post:
      tags: [notify]
      summary: Send a notification
      description: |
        Broadcasts the notification to all connected WebSocket clients.

        If `wait` or `maxWait` is set the notification is marked
        exclusive and the response is held open until a client resolves
        it or the `wait` window elapses, at which point the caller
        receives a pending response and should long-poll
        `/notify/{id}/wait`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Notification"
      responses:
        "200":
          description: Notification was resolved by a client within `wait` seconds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedMessage"
        "202":
          description: |
            Non-exclusive fire-and-forget accepted (no body), OR the
            short `wait` window elapsed and the notification is still
            live — body is a pending response and the `Location` header
            points at the long-poll endpoint.
          headers:
            Location:
              description: Long-poll URL (only present when body is a PendingResponse).
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "410":
          description: Notification's `maxWait` elapsed before any client resolved it.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpiredMessage" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
  /notify/alertmanager:
    post:
      tags: [notify]
      summary: Accept a Prometheus Alertmanager v4 webhook
      description: |
        Accepts the JSON payload that Prometheus Alertmanager posts to
        webhook receivers and translates it into a cross-notifier
        notification, broadcasting on the same path as `POST /notify`.

        Translation summary (full mapping documented in
        `crates/core/src/alertmanager.rs`):
          - `status: "resolved"` → cross-notifier `status: "success"`
          - `status: "firing"` + `commonLabels.severity: "page"` or
            `"critical"` → `status: "error"`
          - `status: "firing"` + `commonLabels.severity: "warning"` →
            `status: "warning"`
          - other firing → `status: "info"`

        The `groupKey` field is hashed into a short stable notification
        id (`am-<11chars>`) so clients can deduplicate Alertmanager's
        periodic re-fires of the same group rather than stacking them.

        Wire format reference:
          https://prometheus.io/docs/alerting/latest/configuration/#webhook_config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: |
                Alertmanager v4 webhook payload. The schema is owned by
                Prometheus; only the fields cross-notifier reads are
                documented here.
              properties:
                version: { type: string, example: "4" }
                groupKey: { type: string }
                status: { type: string, enum: [firing, resolved] }
                receiver: { type: string }
                groupLabels: { type: object, additionalProperties: { type: string } }
                commonLabels: { type: object, additionalProperties: { type: string } }
                commonAnnotations: { type: object, additionalProperties: { type: string } }
                externalURL: { type: string, format: uri }
                alerts:
                  type: array
                  items:
                    type: object
                    properties:
                      status: { type: string, enum: [firing, resolved] }
                      labels: { type: object, additionalProperties: { type: string } }
                      annotations: { type: object, additionalProperties: { type: string } }
                      startsAt: { type: string, format: date-time }
                      endsAt: { type: string, format: date-time }
                      generatorURL: { type: string, format: uri }
                      fingerprint: { type: string }
      responses:
        "202":
          description: |
            Translated and accepted. Broadcasts to WebSocket subscribers
            and (when push is configured) fans out to mobile devices.
        "400":
          description: |
            Payload was not valid Alertmanager v4 JSON, or was so empty
            (no summary, no alertname, no alerts) that no displayable
            notification could be derived.
        "401": { $ref: "#/components/responses/Unauthorized" }
  /notify/{id}/wait:
    get:
      tags: [notify]
      summary: Long-poll for a pending notification's resolution
      description: |
        Continues waiting for a previously-submitted exclusive
        notification whose initial POST returned `202 pending`.
        The connection blocks for up to `timeout` seconds.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: timeout
          in: query
          required: false
          description: Seconds to block this request. Defaults to 25.
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Resolved by a client.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResolvedMessage" }
        "202":
          description: Still pending — caller should poll again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PendingResponse" }
        "410":
          description: Notification expired or is unknown (already cleaned up).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpiredMessage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
  /devices:
    get:
      tags: [devices]
      summary: List registered push devices
      description: |
        Available only when the server is started with a device registry
        (set `-devices-file` or `CROSS_NOTIFIER_DEVICES_FILE`). Returns
        every registered device regardless of platform.
      responses:
        "200":
          description: Device list.
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items: { $ref: "#/components/schemas/Device" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Registry not configured on this server.
    post:
      tags: [devices]
      summary: Register a push device
      description: |
        Upsert by `deviceToken`. Re-registering the same token updates
        the label but preserves the original `registeredAt`. Called by
        the mobile app on every launch so the server can learn fresh
        tokens after reinstalls or iCloud restores.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterDeviceRequest" }
      responses:
        "200":
          description: Registered (or updated).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Registry not configured on this server.
  /devices/{token}:
    delete:
      tags: [devices]
      summary: Unregister a push device
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Unregistered.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Token not found, or registry not configured.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Shared secret, sent as `Authorization: Bearer <secret>`."
  responses:
    BadRequest:
      description: Invalid request body.
    Unauthorized:
      description: Missing or invalid bearer token.
    MethodNotAllowed:
      description: Wrong HTTP method for this endpoint.
  schemas:
    Notification:
      type: object
      required: [source]
      properties:
        id:
          type: string
          description: Optional caller-assigned ID. Server assigns a UUID if omitted on exclusive notifications.
        source:
          type: string
          description: Origin of the notification (logical, not an address).
        title:
          type: string
          description: At least one of `title` or `message` is required.
        message:
          type: string
        status:
          type: string
          enum: [info, success, warning, error]
        iconData:
          type: string
          description: Base64-encoded image.
        iconHref:
          type: string
          format: uri
          description: Server fetches this URL and resizes to 48x48.
        duration:
          type: integer
          description: Seconds before auto-dismiss on the client. 0/omitted = never.
        exclusive:
          type: boolean
          description: Coordinates actions across all clients — first action wins.
        wait:
          type: integer
          minimum: 0
          description: |
            Seconds the POST blocks waiting for a client response. Any
            non-zero value implies `exclusive`. Defaults to 25 when
            unset on an exclusive notification.
        maxWait:
          type: integer
          minimum: 0
          description: |
            Total lifetime in seconds. After this, the server broadcasts
            `expired` over WebSocket and cleans up. Defaults to `wait`.
        actions:
          type: array
          items: { $ref: "#/components/schemas/Action" }
...
    ResolvedMessage:
      type: object
      required: [id, resolvedBy, actionLabel, success]
      properties:
        id: { type: string }
        resolvedBy:
          type: string
          description: X-Client-Name of the client that resolved the notification.
        actionLabel: { type: string }
        success: { type: boolean }
        error:
          type: string
          description: Present when `success` is false.
    ExpiredMessage:
      type: object
      required: [id]
      properties:
        id: { type: string }
    PendingResponse:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [pending]
...
    Device:
      type: object
      required: [token, platform, registeredAt]
      properties:
        token:
          type: string
          description: Opaque push token issued by APNS/FCM.
        label:
          type: string
          description: Human-readable label for the device list UI.
        platform: { $ref: "#/components/schemas/Platform" }
        registeredAt:
          type: string
          format: date-time
        lastPushAt:
          type: string
          format: date-time
          description: Timestamp of the last successful push, if any.
...